Quote str's starting with :, Ruby thinks they're symbols
We had a fatal error with OpenStreetMap.org in Hebrew because some of
our strings started with :. As it turned out this was because Perl's
YAML::Syck will happily emit the string ":foo" as simply :foo, but
Ruby interprets these as symbols, not strings.

See this commit for a hotfix:

    http://github.com/openstreetmap/openstreetmap-website/commit/7f8ed82

With this change anything that's a str, starts with : and is about to
be emitted unquoted will be quoted.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>