Make DumpFile streaming; add DumpInto functions
Previously, DumpFile (YAML and JSON) were implemented by first calling the XS
Dump function to return a string containing the YAML/JSON and then writing that
data to the file/filehandle. This commit changes them to call directly to XS
with a syck_output_handler that will output directly, which therefore avoids
needing to keep the entire YAML/JSON document in memory.

This commit also adds DumpInto functions, which take a reference to a scalar and
append the YAML/JSON data to that scalar, which avoids needing to copy the output
when returning from Dump.

Documentation and tests included.